wip adding torrent blocks

+++ bit_tor/src/encoding.rs
@@ -6,7 +6,7 @@ pub mod encoder;
 #[derive(Clone, Debug, PartialEq)]
 pub enum Value {
     Null,
-    Number(i128),
+    Number(i64),
     // utf-8 str or others
     Bytes(Vec<u8>),
     Array(Vec<Value>),
